Revert "Revert "Normalize WER better""
This reverts commit 186f3580aa9c3bc31da87ac464c6dc3d9b279758.
