+++ Sources/FluidAudio/ASR/TDT/TdtDecoder.swift
@@ -370,7 +370,7 @@ internal struct TdtDecoder {
                     model: jointModel
                 )
 
-                let (tokenLogits, durationLogits) = try splitLogits(
+                let (tokenLogits, _) = try splitLogits(
                     logits, durationElements: config.tdtConfig.durationBins.count)
 
                 let token = argmaxSIMD(tokenLogits)
