--- Sources/FluidAudioCLI/TextNormalizer.swift
@@ -1,232 +1,0 @@
-import Foundation
-import RegexBuilder
-
-/// HuggingFace-compatible text normalizer for ASR evaluation
-/// Matches the normalization used in the Open ASR Leaderboard
-struct TextNormalizer {
-
-    private static let additionalDiacritics: [Character: String] = [
-        "œ": "oe", "Œ": "OE", "ø": "o", "Ø": "O", "æ": "ae", "Æ": "AE",
-        "ß": "ss", "ẞ": "SS", "đ": "d", "Đ": "D", "ð": "d", "Ð": "D",
-        "þ": "th", "Þ": "th", "ł": "l", "Ł": "L",
-    ]
-
-    private static let britishToAmerican: [String: String] = [
-        "colour": "color",
-        "colourless": "colorless",
-        "colouring": "coloring",
-        "colourful": "colorful",
-        "colourist": "colorist",
-        "colourists": "colorists",
-        "counselled": "counseled",
-        "favourite": "favorite",
-        "flavour": "flavor",
-        "favour": "favor",
-        "honour": "honor",
-        "labour": "labor",
-        "neighbour": "neighbor",
-        "centre": "center",
-        "theatre": "theater",
-        "metre": "meter",
-        "litre": "liter",
-        "defence": "defense",
-        "offence": "offense",
-        "licence": "license",
-        "travelling": "traveling",
-        "cancelled": "canceled",
-        "jewellery": "jewelry",
-        "cheque": "check",
-        "grey": "gray",
-        "plough": "plow",
-        "tyre": "tire",
-        "aluminium": "aluminum",
-        "organise": "organize",
-        "realise": "realize",
-        "recognise": "recognize",
-        "apologise": "apologize",
-        "analyse": "analyze",
-        "paralyse": "paralyze",
-        "catalogue": "catalog",
-        "dialogue": "dialog",
-        "programme": "program",
-        "pyjamas": "pajamas",
-        "aeroplane": "airplane",
-        "maths": "math",
-        "mum": "mom",
-            // Add more as needed
-    ]
-
-    /// Normalize text using HuggingFace ASR leaderboard standards
-    /// This matches the normalization used in the official leaderboard evaluation
-    static func normalize(_ text: String) -> String {
-        var normalized = text
-
-        normalized = normalized.lowercased()
-
-        // British to American normalization
-        for (british, american) in britishToAmerican {
-            let pattern = "\\b" + NSRegularExpression.escapedPattern(for: british) + "\\b"
-            normalized = normalized.replacingOccurrences(
-                of: pattern,
-                with: american,
-                options: .regularExpression
-            )
-        }
-
-        let bracketsPattern = try! NSRegularExpression(pattern: "[<\\[].*?[>\\]]", options: [])
-        normalized = bracketsPattern.stringByReplacingMatches(
-            in: normalized,
-            options: [],
-            range: NSRange(location: 0, length: normalized.count),
-            withTemplate: ""
-        )
-
-        let parenthesesPattern = try! NSRegularExpression(pattern: "\\([^)]+?\\)", options: [])
-        normalized = parenthesesPattern.stringByReplacingMatches(
-            in: normalized,
-            options: [],
-            range: NSRange(location: 0, length: normalized.count),
-            withTemplate: ""
-        )
-
-        normalized = normalized.map { char in
-            if let replacement = additionalDiacritics[char] {
-                return replacement
-            }
-            return String(char)
-        }.joined()
-
-        normalized = normalized.replacingOccurrences(of: "$", with: " dollar ")
-        normalized = normalized.replacingOccurrences(of: "&", with: " and ")
-        normalized = normalized.replacingOccurrences(of: "%", with: " percent ")
-
-        let punctuationPattern = try! NSRegularExpression(pattern: "[^\\w\\s']", options: [])
-        normalized = punctuationPattern.stringByReplacingMatches(
-            in: normalized,
-            options: [],
-            range: NSRange(location: 0, length: normalized.count),
-            withTemplate: " "
-        )
-
-        let contractions = [
-            "can't": "cannot",
-            "won't": "will not",
-            "n't": " not",
-            "'re": " are",
-            "'ve": " have",
-            "'ll": " will",
-            "'d": " would",
-            "'m": " am",
-            "it's": "it is",
-            "that's": "that is",
-            "there's": "there is",
-            "here's": "here is",
-            "what's": "what is",
-            "where's": "where is",
-            "who's": "who is",
-            "how's": "how is",
-            "let's": "let us",
-            "i'm": "i am",
-            "you're": "you are",
-            "we're": "we are",
-            "they're": "they are",
-            "you've": "you have",
-            "we've": "we have",
-            "they've": "they have",
-            "i've": "i have",
-            "you'll": "you will",
-            "we'll": "we will",
-            "they'll": "they will",
-            "i'll": "i will",
-            "you'd": "you would",
-            "we'd": "we would",
-            "they'd": "they would",
-            "i'd": "i would",
-            "she's": "she is",
-            "he's": "he is",
-            "she'll": "she will",
-            "he'll": "he will",
-            "she'd": "she would",
-            "he'd": "he would",
-        ]
-
-        for (contraction, expansion) in contractions {
-            normalized = normalized.replacingOccurrences(of: contraction, with: expansion)
-        }
-
-        let abbreviations = [
-            "mr": "mister",
-            "mrs": "misess",
-            "ms": "miss",
-            "dr": "doctor",
-            "prof": "professor",
-            "st": "saint",
-            "jr": "junior",
-            "sr": "senior",
-            "vs": "versus",
-            "inc": "incorporated",
-            "ltd": "limited",
-            "co": "company",
-        ]
-
-        for (abbrev, expansion) in abbreviations {
-            let pattern = "\\b" + abbrev + "\\b"
-            normalized = normalized.replacingOccurrences(
-                of: pattern,
-                with: expansion,
-                options: .regularExpression
-            )
-        }
-
-        let numberWords = [
-            "zero": "0", "one": "1", "two": "2", "three": "3", "four": "4",
-            "five": "5", "six": "6", "seven": "7", "eight": "8", "nine": "9",
-            "ten": "10", "eleven": "11", "twelve": "12", "thirteen": "13",
-            "fourteen": "14", "fifteen": "15", "sixteen": "16", "seventeen": "17",
-            "eighteen": "18", "nineteen": "19", "twenty": "20", "thirty": "30",
-            "forty": "40", "fifty": "50", "sixty": "60", "seventy": "70",
-            "eighty": "80", "ninety": "90", "hundred": "100", "thousand": "1000",
-            "million": "1000000", "billion": "1000000000",
-            "first": "1st", "second": "2nd", "third": "3rd", "fourth": "4th",
-            "fifth": "5th", "sixth": "6th", "seventh": "7th", "eighth": "8th",
-            "ninth": "9th", "tenth": "10th",
-        ]
-
-        for (word, digit) in numberWords {
-            let pattern = "\\b" + word + "\\b"
-            normalized = normalized.replacingOccurrences(
-                of: pattern,
-                with: digit,
-                options: .regularExpression
-            )
-        }
-
-        normalized = normalized.replacingOccurrences(of: "€", with: " euro ")
-        normalized = normalized.replacingOccurrences(of: "£", with: " pound ")
-        normalized = normalized.replacingOccurrences(of: "¥", with: " yen ")
-        normalized = normalized.replacingOccurrences(of: "©", with: " copyright ")
-        normalized = normalized.replacingOccurrences(of: "®", with: " registered ")
-        normalized = normalized.replacingOccurrences(of: "™", with: " trademark ")
-
-        let finalCleanPattern = try! NSRegularExpression(pattern: "[^\\w\\s]", options: [])
-        normalized = finalCleanPattern.stringByReplacingMatches(
-            in: normalized,
-            options: [],
-            range: NSRange(location: 0, length: normalized.count),
-            withTemplate: " "
-        )
-
-        let whitespacePattern = try! NSRegularExpression(pattern: "\\s+", options: [])
-        normalized = whitespacePattern.stringByReplacingMatches(
-            in: normalized,
-            options: [],
-            range: NSRange(location: 0, length: normalized.count),
-            withTemplate: " "
-        )
-
-        normalized = normalized.trimmingCharacters(in: .whitespacesAndNewlines)
-
-        return normalized
-    }
-
-}
